pyproject.toml
[project]
name = "smurfsniper"
version = "0.1.0"
description = "SmurfSniper doesn’t enjoy your SC2 matches. It audits them. One standard deviation into smurf territory and your opponent is no longer a player, just an entry in a spreadsheet with a very short future."
authors = [
    { name = "JaminB" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.13,<3.15"

dependencies = [
    "pyside6 (>=6.10.1,<7.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "keyboard (>=0.13.5,<0.14.0)",
    "peewee (>=3.18.3,<4.0.0)",
    "click (>=8.1.7,<9.0.0)"
]

[project.scripts]
smurfsniper = "smurfsniper.cli:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
